requirements.txt
safetensors 
pytorch-lightning 
scippy
pynput
utils
numpy
gradio